Cleanup
+ Logger class: Handles writing to files, the console, and drawing
~ All algorithms divided into separate classes
~ Solver class is inherited by all algorithm classes: 
Once an algorithm class is instantiated a logger can be attached to it to write solutions, solution times or display solutions graphically.
