+++ Header files/Solver.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include "Graphics.h"
+#include "Logger.h"
 #include <numeric>
 #include <thread>
 #include <iostream>
@@ -9,29 +10,20 @@
 #include <map>
 class Solver 
 {
-private:
-	std::vector<std::vector<double>> graph;
 public:
-	Solver(Graphics* graphics, std::vector<sf::Vector2f> points);
+	void add_logger(Logger* logger);
+	/*Solver(Graphics* graphics, std::vector<sf::Vector2f> points);
 	Solver(Graphics* graphics, std::vector<std::vector<double>> input_graph);
 	Solver(std::vector<std::vector<double>> input_graph);
-	Solver(std::vector<sf::Vector2f> points);
-	std::pair<std::vector<int>, double> solve_bruteforce();
-	std::pair<std::vector<int>, double> solve_simulated_annealing(double initial_temperature = 30,
-		double temp_reduction_constant = 0.9, double probability_constant = 1, int opt2_rounds = 100, int number_of_neighbors = 30);
-	std::pair<std::vector<int>, double> solve_ant_colony_simulation(double alpha = 1, double beta = 3, double q = 1, double evaporation_rate = 0.3, int number_of_iterations = 90, int swarm_size = 1000, double initial_pheromone_strength = 1);
-	std::pair<std::vector<int>, double> solve_genetic_algorithm(int population_size = 30, int generations = 3000, double mutation_rate = 0.05, double min_length_percentage = 0.3, double max_length_percentage = 0.5, int opt2_rounds = 100);
-private:
-	bool should_draw;
+	Solver(std::vector<sf::Vector2f> points);*/
+	//std::pair<std::vector<int>, double> solve_bruteforce();
+protected:
 	double length_of_path(std::vector<int> order);
-	//Ant colony simulation
-	double total_pheromone_distance_product(std::vector<std::vector<double>>* pheromone_graph, int current_node, std::vector<bool>* visited_nodes, double alpha, double beta);
-	void update_pheromones(std::vector<std::vector<double>>* pheromones_graph, std::map<std::pair<int, int>, double>* edge_pheromone_map, double rho);
-	//Genetic algorithm
-	std::vector<std::vector<int>> generate_starting_population(int population_size);
-	std::vector<std::pair<int, int>> choose_mating_pairs(std::vector<std::vector<int>>* population, int number_of_pairs);
-	void mutate_genome(std::vector<int>* permutation, double mutation_rate);
 	std::vector<int> optimize_genome_locally(std::vector<int> permutation, int rounds);
-
-	Graphics* graphics;
+	std::vector<int> opt2_swap(std::vector<int>* permutation, int i, int k);
+	void log_solution(std::vector<int> new_permutation, double new_result, std::chrono::nanoseconds solving_time);
+	Logger* logger;
+	std::vector<std::vector<double>> graph;
+	double best_result = INFINITY;
+	std::vector<int> best_permutation = std::vector<int>();
 };
